Closes issue #32 by adding a text editor within VisualGit
* Added incomplete text editor feature

Only added an initial text editor that doesn't close back after the user is done editing a file. This is merely a test of the text editing functionality. Whole team present.

* Improved text edit flow

It now checks if text view is open before edit is opened. Can discard changes and also alert the user if trying to open view  text mode while in edit mode

* Moved css elements to scss for text edit feature

Previously scss was not working, so we made changes to css for text edit features. Since now it works we transferred it to scss.

* Fixed display problem with text editor 

Text editor was partially displayed when it's supposed to be hidden completely, that got fixed.

Made some variable and function name changes to make code more understandable.

* Added confirmation for text editor exit points

- Corrected the flow problem of discarding changes without warning when user tries to right click and edit another file when editing some file
- Added confirmation dialog for exiting editing mode instead of just having an information dialog
- Fixed potential bug of using OR operator with size comparison with 0 and '' in index.html
- Made some function/variable names more descriptive

* Added file name to text editor

This is to help better orientate the user

* Ensured that text editor always start with first line of text

- Text editor will always scroll to the top.
- Added tooltip to the title of the file being edited