Closes issue #122 Disable empty commits
This solution checks if there are any modified files waiting to be committed in the file.panel component. If there are no changes waiting to be committed, the [Commit] button is disabled and the opacity of the button is reduced to indicate that it cannot be used.
Everytime git.ts#displayModifiedFiles() is called, it checks if it is valid/invalid to enable the [Commit] button at the end of that function.